refactor: change models name in properties folder
