+++ internal/properties/properties.handler.go
@@ -1,4 +1,4 @@
-package property
+package properties
 
 import (
 	"github.com/brain-flowing-company/pprp-backend/internal/models"
@@ -28,14 +28,14 @@ func NewHandler(service Service) Handler {
 // @description Get property by its id
 // @tags        property
 // @produce     json
-// @success     200	{object} models.Property
+// @success     200	{object} models.Properties
 // @failure     400 {object} models.ErrorResponse "Invalid property id"
 // @failure     404 {object} models.ErrorResponse "Property id not found"
 // @failure     500 {object} models.ErrorResponse
 func (h *handlerImpl) GetPropertyById(c *fiber.Ctx) error {
 	propertyId := c.Params("propertyId")
 
-	property := models.Property{}
+	property := models.Properties{}
 	err := h.service.GetPropertyById(&property, propertyId)
 	if err != nil {
 		return utils.ResponseError(c, err)
@@ -50,7 +50,7 @@ func (h *handlerImpl) GetPropertyById(c *fiber.Ctx) error {
 // @tags        property
 // @produce     json
 // @param       query query string true "Search query"
-// @success     200	{object} []models.Property
+// @success     200	{object} []models.Properties
 // @failure     500 {object} models.ErrorResponse
 func (h *handlerImpl) GetOrSearchProperties(c *fiber.Ctx) error {
 	query := c.Query("query")
@@ -62,7 +62,7 @@ func (h *handlerImpl) GetOrSearchProperties(c *fiber.Ctx) error {
 }
 
 func (h *handlerImpl) GetAllProperties(c *fiber.Ctx) error {
-	properties := []models.Property{}
+	properties := []models.Properties{}
 	err := h.service.GetAllProperties(&properties)
 	if err != nil {
 		return utils.ResponseError(c, err)
@@ -74,7 +74,7 @@ func (h *handlerImpl) GetAllProperties(c *fiber.Ctx) error {
 func (h *handlerImpl) SeachProperties(c *fiber.Ctx) error {
 	query := c.Query("query")
 
-	properties := []models.Property{}
+	properties := []models.Properties{}
 	err := h.service.SearchProperties(&properties, query)
 	if err != nil {
 		return utils.ResponseError(c, err)
